requirements.txt
attrs~=21.2.0
python-dotenv~=0.18.0
requests~=2.26.0
pytest